requirements.txt
appdirs==1.4.0
beautifulsoup4==4.5.3
boto3==1.4.4
botocore==1.5.10
bs4==0.0.1
datadiff==2.0.0
docutils==0.13.1
future==0.16.0
futures==3.0.5
Jinja2==2.8.1
jmespath==0.9.1
lamvery==0.18.1
MarkupSafe==0.23
oauthlib==2.0.1
packaging==16.8
parsedatetime==2.2
Pygments==2.2.0
pyparsing==2.1.10
python-dateutil==2.6.0
PyYAML==3.12
requests==2.13.0
requests-oauthlib==0.7.0
s3transfer==0.1.10
six==1.10.0
termcolor==1.1.0
twython==3.4.0
